fix(earnings_dag.py) testing new int conversion

+++ plugins/earnings_data_pull_step.py
@@ -480,7 +480,7 @@ class SQLMinuteToDailyEquityPrices_2_0(DataReaderClass):
         data = data.groupby(['ticker', 'date']).mean()[
             ['open', 'high', 'low', 'close', 'volume', 'dcm_security_id']].reset_index().sort_values(by=['date'])
 
-        data['dcm_security_id'] = data['dcm_security_id'].astype(int)
+        #data['dcm_security_id'] = data['dcm_security_id'].astype(int)
 
         return data
